Fix API limit issue
When fetching all tasks ensure that the existing AWS Client is passed to the new task object and re-used. Otherwise if you have lots of tasks when you run under an ECS task you will hit the rate limit asking for new credentials for each new client that is created. This rate limit on ECS appears to be a burst of 60 and steady rate of 40.